+++ src/main/java/com/airbnb/bookingservice/LocationService.java
@@ -3,5 +3,5 @@ package com.airbnb.bookingservice;
 import org.springframework.stereotype.Service;
 
 public interface LocationService {
-    Location getLocationDetails(int id);
+    Location getLocationDetails(long id);
 }
